refactor: 시간에 대한 처리를 controller를 따로 두어서 처리
